JPFL 15/03

+++ DRRCore.Domain.Entities/SqlCoreContext/AgentPrice.cs
@@ -17,31 +17,31 @@ public partial class AgentPrice
 
     public int? IdCurrency { get; set; }
 
-    public int? PriceT1 { get; set; }
+    public decimal? PriceT1 { get; set; }
 
     public int? DayT1 { get; set; }
 
-    public int? PriceT2 { get; set; }
+    public decimal? PriceT2 { get; set; }
 
     public int? DayT2 { get; set; }
 
-    public int? PriceT3 { get; set; }
+    public decimal? PriceT3 { get; set; }
 
     public int? DayT3 { get; set; }
 
-    public int? PricePn { get; set; }
+    public decimal? PricePn { get; set; }
 
     public int? DayPn { get; set; }
 
-    public int? PriceBd { get; set; }
+    public decimal? PriceBd { get; set; }
 
     public int? DayBd { get; set; }
 
-    public int? PriceRp { get; set; }
+    public decimal? PriceRp { get; set; }
 
     public int? DayRp { get; set; }
 
-    public int? PriceCr { get; set; }
+    public decimal? PriceCr { get; set; }
 
     public int? DayCr { get; set; }
